move requestData creation to view
